backend_v2.5.0

+++ echoq/src/main/java/org/echoq/controller/UserController.java
@@ -6,6 +6,7 @@ import org.echoq.entity.User;
 import org.echoq.service.IUserService;
 import org.echoq.util.Result;
 import org.echoq.util.ResultToken;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.web.bind.annotation.*;
 
 import org.springframework.web.multipart.MultipartFile;
@@ -33,8 +34,11 @@ public class UserController {
     @Resource
     IUserService service;
 
+    @Resource
+    RedisTemplate<String, Object> template;
+
     @PostMapping("/signup")
-    public ResultToken<Integer> signup(@RequestBody User user, HttpSession session)
+    public ResultToken<Integer> signup(@RequestBody User user)
     {
         String username = user.getUsername();
         int replicates = service.selectUsername(username);
@@ -44,7 +48,8 @@ public class UserController {
         if (response == 1) {
             String uuid = UUID.randomUUID().toString();
             int id = service.selectUser(username);
-            session.setAttribute(uuid, id);
+            template.opsForValue().set(uuid,id);
+            // session.setAttribute(uuid, id);
             return new ResultToken<>(1, 200, uuid);
         }
         else
@@ -52,7 +57,7 @@ public class UserController {
     }
 
     @PostMapping("/signin")
-    public ResultToken<Integer> signin(@RequestBody User user, HttpSession session)
+    public ResultToken<Integer> signin(@RequestBody User user)
     {
         String username = user.getUsername();
         String password = user.getPassword();
@@ -60,17 +65,26 @@ public class UserController {
         int userId = service.selectUser(username);
         if (c == 1) {
             String uuid = UUID.randomUUID().toString();
-            session.setAttribute(uuid, userId);
+            template.opsForValue().set(uuid,userId);
+            // session.setAttribute(uuid, userId);
             return new ResultToken<>(1, 200, uuid);
         }else
             return new ResultToken<>(-1, 400, null);
     }
 
-    @GetMapping("/getInfo")
-    public Result<User> getInfo(HttpServletRequest request, HttpSession session)
+    @PostMapping("/signout")
+    public Result<Integer> signout(HttpServletRequest request)
     {
         String key = request.getHeader("XXX-SToken");
-        int id = (int) session.getAttribute(key);
+        template.delete(key);
+        return new Result<>(1, 200);
+    }
+
+    @GetMapping("/getInfo")
+    public Result<User> getInfo(HttpServletRequest request)
+    {
+        String key = request.getHeader("XXX-SToken");
+        Integer id = (Integer) template.opsForValue().get(key);
         User user = service.selectUserInfo(id);
         if (user.getUsername().equals("") || user.getUsername() == null)
         {
@@ -82,16 +96,18 @@ public class UserController {
     }
 
     @PostMapping("/updateInfo")
-    public Result<Integer> update(String name, String password, String whatsup, String color, HttpSession session)
+    public Result<Integer> update(@RequestBody User user, HttpServletRequest request)
     {
-        int id = (int) session.getAttribute("loginId");
-        User usr = new User();
+        String key = request.getHeader("XXX-SToken");
+        Integer id = (Integer) template.opsForValue().get(key);
+        user.setUserid(id);
+        /*User usr = new User();
         usr.setUserid(id);
         usr.setName(name);
         usr.setPassword(password);
         usr.setWhatsup(whatsup);
-        usr.setColor(color);
-        int res = service.updateUserInfo(usr);
+        usr.setColor(color);*/
+        int res = service.updateUserInfo(user);
         if (res == 1)
             return new Result<>(1, 200);
         else
@@ -99,15 +115,16 @@ public class UserController {
     }
 
     @GetMapping("/selectQuestions")
-    public Result<List<Questions>> selectQuestions(String condition, HttpSession session)
+    public Result<List<Questions>> selectQuestions(String condition, HttpServletRequest request)
     {
-        int id = (int) session.getAttribute("loginId");
+        String key = request.getHeader("XXX-SToken");
+        Integer id = (Integer) template.opsForValue().get(key);
         List<Questions> questions = service.selectQuestionConditional(id, condition);
         return new Result<>(questions, 200);
     }
 
     @GetMapping("/getQ/{username}")
-    public Result<List<Questions>> selectQuestionsInvisible(@PathVariable String username, String condition)
+    public Result<List<Questions>> selectQuestionsVisible(@PathVariable String username, String condition)
     {
         List<Questions> questions = service.selectQuestionConditionalInvisible(username, condition);
         return new Result<>(questions, 200);
@@ -115,11 +132,12 @@ public class UserController {
 
 
     @PostMapping("updateAvatar")
-    public Result<Integer> updateAvatar(MultipartFile img, HttpSession session) throws IOException {
-        int id = (int) session.getAttribute("loginId");
+    public Result<Integer> updateAvatar(MultipartFile img, HttpServletRequest request) throws IOException {
+        String key = request.getHeader("XXX-SToken");
+        Integer id = (Integer) template.opsForValue().get(key);
         File file = new File("src/main/resources/targetFile.tmp");
         img.transferTo(file);
-        int res = service.updateAvatar(file);
+        int res = service.updateAvatar(file, id);
         if (res == 1)
             return new Result<>(1, 200);
         else
@@ -127,33 +145,50 @@ public class UserController {
     }
 
     @PostMapping("respondQuestion")
-    public Result<Integer> respondQuestion(int questionId, String response, boolean visibility)
+    public Result<Integer> respondQuestion(int questionId, String response, boolean visibility, HttpServletRequest request)
     {
-        Questions q = new Questions();
-        q.setQuestionid(questionId);
-        q.setResponse(response);
-        q.setVisibility(visibility);
-        int res = service.respondQuestion(q);
-        if (res == 1)
-            return new Result<>(1, 200);
-        else
-            return new Result<>(-1, 400);
+        String key = request.getHeader("XXX-SToken");
+        Integer id = (Integer) template.opsForValue().get(key);
+        int confId = service.confirmQuestionBelonging(questionId);
+        if (id == confId)
+        {
+            Questions q = new Questions();
+            q.setQuestionid(questionId);
+            q.setResponse(response);
+            q.setVisibility(visibility);
+            int res = service.respondQuestion(q);
+            if (res == 1)
+                return new Result<>(1, 200);
+            else
+                return new Result<>(-1, 400);
+        }else
+        {
+            return new Result<>(-1, 401);
+        }
     }
 
     @PostMapping("/updateVisibility")
-    public Result<Integer> updateVisibility(int questionId, boolean visibility)
+    public Result<Integer> updateVisibility(int questionId, boolean visibility, HttpServletRequest request)
     {
-        int res = service.updateVisibility(visibility, questionId);
-        if (res == 1)
-            return new Result<>(1, 200);
-        else
-            return new Result<>(-1, 400);
+        String key = request.getHeader("XXX-SToken");
+        Integer id = (Integer) template.opsForValue().get(key);
+        int confId = service.confirmQuestionBelonging(questionId);
+        if (id == confId)
+        {
+            int res = service.updateVisibility(visibility, questionId);
+            if (res == 1)
+                return new Result<>(1, 200);
+            else
+                return new Result<>(-1, 400);
+        }else
+            return new Result<>(-1, 401);
     }
 
     @DeleteMapping("/deleteUser")
-    public Result<Integer> deleteUser(HttpSession session)
+    public Result<Integer> deleteUser(HttpServletRequest request)
     {
-        int id = (int) session.getAttribute("loginId");
+        String key = request.getHeader("XXX-SToken");
+        Integer id = (Integer) template.opsForValue().get(key);
         int res = service.deleteUser(id);
         if (res == 1)
             return new Result<>(1, 200);
@@ -162,35 +197,50 @@ public class UserController {
     }
 
     @DeleteMapping("/deleteQuestion")
-    public Result<Integer> deleteQuestion(int questionId)
+    public Result<Integer> deleteQuestion(int questionId, HttpServletRequest request)
     {
-        int res = service.deleteQuestion(questionId);
-        if (res == 1)
-            return new Result<>(1, 200);
-        else
-            return new Result<>(-1, 400);
+        String key = request.getHeader("XXX-SToken");
+        Integer id = (Integer) template.opsForValue().get(key);
+        int confId = service.confirmQuestionBelonging(questionId);
+        if (id == confId)
+        {
+            int res = service.deleteQuestion(questionId);
+            if (res == 1)
+                return new Result<>(1, 200);
+            else
+                return new Result<>(-1, 400);
+        }else
+            return new Result<>(-1, 401);
     }
 
     @DeleteMapping("/deleteQuestions")
-    public Result<Integer> deleteQuestions(int[] questionId)
+    public Result<Integer> deleteQuestions(int[] questionId, HttpServletRequest request)
     {
-        boolean isSucceed = true;
-        for (int i = 0; i < questionId.length; i ++)
+        String key = request.getHeader("XXX-SToken");
+        Integer id = (Integer) template.opsForValue().get(key);
+        int confId = service.confirmQuestionBelonging(questionId[0]);
+        if (id == confId)
         {
-            int res = service.deleteQuestion(questionId[i]);
-            if (res != 1)
-                isSucceed = false;
-        }
-        if (isSucceed)
-            return new Result<>(1, 200);
-        else
-            return new Result<>(-1, 400);
+            boolean isSucceed = true;
+            for (int i = 0; i < questionId.length; i ++)
+            {
+                int res = service.deleteQuestion(questionId[i]);
+                if (res != 1)
+                    isSucceed = false;
+            }
+            if (isSucceed)
+                return new Result<>(1, 200);
+            else
+                return new Result<>(-1, 400);
+        }else
+            return new Result<>(-1, 401);
     }
 
     @GetMapping("/searchQuestion")
-    public Result<List<Questions>> searchQuestion(String searchContent, HttpSession session) {
-        int userId = (int) session.getAttribute("loginId");
-        return new Result<>(service.searchQuestion(userId, searchContent), 200);
+    public Result<List<Questions>> searchQuestion(String searchContent, HttpServletRequest request) {
+        String key = request.getHeader("XXX-SToken");
+        Integer id = (Integer) template.opsForValue().get(key);
+        return new Result<>(service.searchQuestion(id, searchContent), 200);
     }
 }
